feat(stats): remove book-list cap with disk-paged entries
The stats book list held a fixed 9-entry array, and beginSession reused
the last slot once full -- so opening a 10th book silently overwrote an
existing book's accumulated stats rather than just truncating the list.

Uncapping the array was not viable: BookStatEntry is 488 bytes and the
array was a member of the resident singleton, so a 500-book library
would have cost 238 KB of the 380 KB budget. Instead each book now
carries a 16-byte resident BookStatIndexEntry holding the fields the
whole-library sweeps need (count, filter, sort, identity hash), and the
full entries stay in stats.bin, faulted in through a 4-slot LRU cache
for only the rows actually drawn. The list already paged at 3 visible
rows, so the working set is tiny.

Entries never move on disk: the index owns display order, so sorting is
a resident stable_sort with no SD I/O, and session writes seek to one
record instead of rewriting the file.

Static DRAM drops 1,904 bytes (110,156 -> 108,252) while the cap goes
away entirely, bounded now only by a 4,000-entry corrupt-header guard.

stats.bin v8 -> v9 appends bookCountTotal at the struct tail, keeping
every v4-v8 prefix byte-identical; v4-v8 files migrate in place.

Verified: clang-format, cppcheck, pio run, 148/148 host tests, and the
simulator builds and boots. Not verified on hardware -- a real device
pass across all 4 orientations with an existing stats.bin is still
needed, particularly the v8 -> v9 migration and the >9-book paging.

+++ src/activities/stats/DetailedStatsActivity.cpp
@@ -29,7 +29,7 @@ void fmtDuration(uint32_t ms, char* buf, size_t n) {
 }
 }  // namespace
 
-DetailedStatsActivity::DetailedStatsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, uint8_t bookIndex)
+DetailedStatsActivity::DetailedStatsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, uint16_t bookIndex)
     : Activity("DetailedStats", renderer, mappedInput), _bookIndex(bookIndex) {}
 
 void DetailedStatsActivity::onEnter() {
