+++ src/activities/stats/DetailedStatsActivity.h
@@ -8,7 +8,7 @@ struct BookStatEntry;
 
 class DetailedStatsActivity final : public Activity {
  public:
-  DetailedStatsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, uint8_t bookIndex);
+  DetailedStatsActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, uint16_t bookIndex);
 
   void onEnter() override;
   void onExit() override;
@@ -20,5 +20,5 @@ class DetailedStatsActivity final : public Activity {
   void drawCover(const BookStatEntry& book, int x, int y, int w, int h) const;
   void drawCoverPlaceholder(int x, int y, int w, int h, const char* title) const;
 
-  uint8_t _bookIndex;
+  uint16_t _bookIndex;
 };
